Move candela source code into dedicated directory
