+++ src/candela/core/VisualizationComponent.js
