+++ src/candela/core/index.js
